refactored duplicated code
