+++ client/src/components/NeedsValidation.jsx
@@ -1,0 +1,19 @@
+function NeedsValidation() {
+  const forms = document.getElementsByClassName('needs-validation');
+  // Loop over them and prevent submission
+  Array.prototype.filter.call(forms, (form) => {
+    form.addEventListener(
+      'submit',
+      (event) => {
+        if (form.checkValidity() === false) {
+          event.preventDefault();
+          event.stopPropagation();
+        }
+        form.classList.add('was-validated');
+      },
+      false,
+    );
+  });
+}
+
+export default NeedsValidation;
